fix odom_publisher dependency
