--- common/common_python/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/common_python
-[install]
-install-scripts=$base/lib/common_python
