+++ odometry_publisher/package.xml
@@ -12,9 +12,7 @@
   <depend>tf2</depend>
   <depend>tf2_ros</depend>
   <depend>sensor_msgs</depend>
-  <depend>can_msgs</depend>
   <depend>nav_msgs</depend>
-  <depend>common_cpp</depend>
   <depend>socketcan_interface_msg</depend>
   
   <test_depend>ament_lint_auto</test_depend>
