+++ odometry_publisher/src/common.cpp
@@ -1,6 +1,18 @@
-#include "to_geometry_msgs.hpp"
+#include "common.hpp"
 
 namespace aiformula {
+namespace odometry_publisher {
+
+void broadcastTf(std::unique_ptr<tf2_ros::TransformBroadcaster>& odometry_br, const nav_msgs::msg::Odometry& odom) {
+    geometry_msgs::msg::TransformStamped ts;
+    ts.header = odom.header;
+    ts.child_frame_id = odom.child_frame_id;
+    ts.transform.translation = toVector3Msg(odom.pose.pose.position);
+    ts.transform.rotation = odom.pose.pose.orientation;
+    odometry_br->sendTransform(ts);
+}
+
+}  // namespace odometry_publisher
 
 geometry_msgs::msg::Point toPointMsg(const double& x, const double& y, const double& z) {
     geometry_msgs::msg::Point point;
