+++ socketcan/socketcan_interface/package.xml
@@ -11,6 +11,7 @@
 
   <depend>rclcpp</depend>
   <depend>socketcan_interface_msg</depend>
+  <depend>libboost-dev</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
