+++ zedx_image_publisher/setup.cfg
